requirements.txt
MouseInfo==0.1.3
packaging==24.2
pillow==11.0.0
psutil==6.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytweening==1.2.0
pywin32==308
six==1.17.0
